+++ LayoutTests/http/wpt/webrtc/peer-connection-color-space-expected.txt
@@ -1,0 +1,5 @@
+
+PASS Per-track colorSpace round-trip and swap (VP9, in-bitstream tag)
+PASS Per-track colorSpace round-trip and swap (VP9, in-bitstream tag stripped, RTP metadata path)
+PASS Per-track colorSpace round-trip and swap (AV1, in-bitstream tag)
+
