+++ LayoutTests/platform/glib/TestExpectations
@@ -1924,6 +1924,7 @@ imported/w3c/web-platform-tests/webcodecs/full-cycle-test.https.any.worker.html?
 
 http/tests/webcodecs/vp9-hdr.html [ Failure ]
 http/tests/webcodecs/av1-sdr.html [ Failure ]
+http/wpt/webrtc/peer-connection-color-space.html [ Failure ]
 
 # Test decoding a corrupt frame after reset in a flush callback assert_unreached: Should have rejected: undefined Reached unreachable code
 imported/w3c/web-platform-tests/webcodecs/videoDecoder-codec-specific.https.any.html?av1 [ Failure ]
