+++ Source/WebCore/platform/video-codecs/cocoa/WebRTCVideoDecoder.h
@@ -51,6 +51,8 @@ public:
     virtual int32_t decodeFrame(int64_t timeStamp, std::span<const uint8_t>) = 0;
     virtual void setFrameSize(uint16_t width, uint16_t height) = 0;
 
+    WEBCORE_EXPORT void setColorSpaceOverride(std::optional<PlatformVideoColorSpace>&&);
+
 protected:
     explicit WebRTCVideoDecoder(std::optional<PlatformVideoColorSpace>&& colorSpaceOverride)
         : m_colorSpaceOverride(WTF::move(colorSpaceOverride))
@@ -60,6 +62,8 @@ protected:
     const std::optional<PlatformVideoColorSpace>& colorSpaceOverride() const { return m_colorSpaceOverride; }
 
 private:
+    virtual void colorSpaceOverrideChanged() { };
+
     std::optional<PlatformVideoColorSpace> m_colorSpaceOverride;
 };
