+++ Source/WebKit/GPUProcess/webrtc/LibWebRTCCodecsProxy.messages.in
@@ -33,6 +33,7 @@ messages -> LibWebRTCCodecsProxy {
     ReleaseDecoder(WebKit::VideoDecoderIdentifier id)
     FlushDecoder(WebKit::VideoDecoderIdentifier id) -> ()
     SetDecoderFormatDescription(WebKit::VideoDecoderIdentifier id, std::span<const uint8_t> description, uint16_t width, uint16_t height)
+    SetDecoderColorSpaceOverride(WebKit::VideoDecoderIdentifier id, struct std::optional<WebCore::PlatformVideoColorSpace> colorSpaceOverride)
     DecodeFrame(WebKit::VideoDecoderIdentifier id, int64_t timeStamp, std::span<const uint8_t> data) -> (bool success)
     SetFrameSize(WebKit::VideoDecoderIdentifier id, uint16_t width, uint16_t height)
